+++ models/BaseUserRelationModel.php
@@ -5,9 +5,9 @@
  * | | / // // ___//_  _//   ||  __||_   _|
  * | |/ // /(__  )  / / / /| || |     | |
  * |___//_//____/  /_/ /_/ |_||_|     |_|
- * @link http://vistart.name/
+ * @link https://vistart.name/
  * @copyright Copyright (c) 2016 vistart
- * @license http://vistart.name/license/
+ * @license https://vistart.name/license/
  */
 
 namespace vistart\Models\models;
@@ -19,6 +19,12 @@ use vistart\Models\queries\BaseUserRelationQuery;
  * This abstract class helps you build user relation model.
  * If you want to use group feature, the BaseUserRelationGroup's extended class
  * must be used coordinately.
+ * Basic usage:
+ * ~~~php
+ * class Relation extends BaseUserRelationModel {
+ * 
+ * }
+ * ~~~
  * @version 2.0
  * @author vistart <i@vistart.name>
  */
@@ -26,7 +32,6 @@ abstract class BaseUserRelationModel extends BaseBlameableModel
 {
     use UserRelationTrait;
 
-    public $idAttribute = false;
     public $confirmationAttribute = false;
     public $contentAttribute = false;
     public $updatedByAttribute = false;