+++ traits/UserTrait.php
@@ -5,15 +5,20 @@
  * | | / // // ___//_  _//   ||  __||_   _|
  * | |/ // /(__  )  / / / /| || |     | |
  * |___//_//____/  /_/ /_/ |_||_|     |_|
- * @link http://vistart.name/
+ * @link https://vistart.name/
  * @copyright Copyright (c) 2016 vistart
- * @license http://vistart.name/license/
+ * @license https://vistart.name/license/
  */
 
 namespace vistart\Models\traits;
 
 /**
  * Assemble PasswordTrait, RegistrationTrait and IdentityTrait into UserTrait.
+ * This trait can only be used in the class extended from [[BaseEntityModel]],
+ * [[BaseMongoEntityModel]], [[BaseRedisEntityModel]], or any other classes used
+ * [[EntityTrait]].
+ * This trait implements two methods `create()` and `findOneOrCreate()`.
+ * Please read the notes of methods and used traits for further detailed usage.
  *
  * @version 2.0
  * @author vistart <i@vistart.name>
@@ -25,16 +30,22 @@ trait UserTrait
         IdentityTrait;
 
     /**
-     * Create new entity model associated with current user.
+     * Create new entity model associated with current user. The model to be created
+     * must be extended from [[BaseBlameableModel]], [[BaseMongoBlameableModel]],
+     * [[BaseRedisBlameableModel]], or any other classes used [[BlameableTrait]].
      * if $config does not specify `userClass` property, self will be assigned to.
      * @param string $className Full qualified class name.
      * @param array $config name-value pairs that will be used to initialize
      * the object properties.
      * @param boolean $loadDefault Determines whether loading default values
      * after entity model created.
+     * Notice! The [[\yii\mongodb\ActiveRecord]] and [[\yii\redis\ActiveRecord]]
+     * does not support load default value. If you want to assign properties
+     * with default values, please define the `default` rule(s) for properties in
+     * `rules()` method and return them by yourself.
      * @param boolean $skipIfSet whether existing value should be preserved.
      * This will only set defaults for attributes that are `null`.
-     * @return $className
+     * @return [[$className]] new model created with specified configuration.
      */
     public function create($className, $config = [], $loadDefault = true, $skipIfSet = true)
     {
@@ -54,11 +65,18 @@ trait UserTrait
     }
 
     /**
-     * Find existed or create new model.
-     * @param string $className
-     * @param array $condition
-     * @param array $config
-     * @return $className
+     * Find existed, or create new model.
+     * If model to be found doesn't exist, and $config is null, the parameter
+     * `$condition` will be regarded as properties of new model.
+     * If you want to know whether the returned model is new model, please check 
+     * the return value of `getIsNewRecord()` method.
+     * @param string $className Full qualified class name.
+     * @param array $condition Search condition, or properties if not found and
+     * $config is null.
+     * @param array $config new model's configuration array. If you specify this
+     * parameter, the $condition will be skipped when created one.
+     * @return [[$className]] the existed model, or new model created by specified
+     * condition or configuration.
      */
     public function findOneOrCreate($className, $condition = [], $config = null)
     {